change QuorumSize from type int to type byte

+++ src/quorum/quorum.go
@@ -13,7 +13,7 @@ import (
 )
 
 const (
-	QuorumSize     int    = 4        // max siblings per quorum
+	QuorumSize     byte   = 4        // max siblings per quorum
 	AtomSize       int    = 4096     // in bytes
 	AtomsPerQuorum int    = 16777216 // 64GB
 	AtomsPerSector uint16 = 200      // more causes DOS problems, is fixable. Final value likely to be 2^9-2^12
